chore: remove abstract database definition
As the abstract class is only used once. The primary reason for having an
abstract class is to encapsulate common behavior that can be shared
across multiple derived classes. There is only one derived class,
so that benefit is lost.
